refactor: 重構todoList
