+++ app/todolist/components/deleteButton.tsx
@@ -1,7 +1,12 @@
 "use client"; // This is a client component 👈🏽
 import React from 'react';
 
-const DeleteButton = ({ id, handleDelete }: DeleteButtonProps) => {
+interface DeleteButtonProps {
+    id: number;
+    handleDelete: DeleteTodo;
+}
+
+const DeleteButton: React.FC<DeleteButtonProps> = ({ id, handleDelete }) => {
     return (
         <button
             className="ml-auto bg-red-500 text-white px-4 py-2 rounded flex items-center justify-center"
